debug: reuse existing textures

+++ source/emulator/debug/cpuRenderer.hpp
@@ -52,10 +52,10 @@ private:
 	Font _font {nullptr};
 	int _font_size {24};
 
-	int _width {513};
-	int _row_height {30};
-	int _vertical_padding {20};
-	int _height {(this->_row_height * 5) + (this->_vertical_padding * 2)};
+	uint64_t _width {513};
+	uint64_t _row_height {30};
+	uint64_t _vertical_padding {20};
+	uint64_t _height {(this->_row_height * 5) + (this->_vertical_padding * 2)};
 	Texture _texture {nullptr};
 
 	void loadRenderer();
